requirements.txt
async-timeout==3.0.1
discord.py==1.7.3
PyNaCl==1.3.0
youtube-dl==2021.12.17
requests~=2.28.1
python-dateutil~=2.8.2